feat: setup mocked config
