+++ react/lib/altpayment/index.ts
@@ -1,5 +1,7 @@
 import { Socket } from "socket.io-client"
-import { SideshiftCoin, SideshiftError, SideshiftPair, SideshiftQuote, SideshiftShift } from "./sideshift"
+import { MockedPaymentClient } from "./mocked"
+import { SideshiftClient, SideshiftCoin, SideshiftError, SideshiftPair, SideshiftQuote, SideshiftShift } from "./sideshift"
+import config from '../config.json'
 
 export const SOCKET_MESSAGES = {
   GET_ALTPAYMENT_RATE: 'get-altpayment-rate',
@@ -11,6 +13,7 @@ export const SOCKET_MESSAGES = {
   ERROR_WHEN_CREATING_SHIFT: 'shift-creation-error'
 }
 
+export type AltpaymentClientOptions = 'sideshift' | 'mocked'
 
 export type AltpaymentCoin = SideshiftCoin
 export type AltpaymentQuote = SideshiftQuote
@@ -18,41 +21,17 @@ export type AltpaymentPair = SideshiftPair
 export type AltpaymentPayment = SideshiftShift
 export type AltpaymentError = SideshiftError
 
-interface AltpaymentListenerParams {
-  addressType: string
-  socket: Socket
-  setCoins: Function
-  setCoinPair: Function
-  setLoadingPair: Function
-  setAltpaymentShift: Function
-  setLoadingShift: Function
-  setAltpaymentError: Function
-}
-
 export interface AltpaymentClient {
   getPaymentStatus: (id: string) => Promise<AltpaymentPayment>
 }
 
-export const altpaymentListener = (params: AltpaymentListenerParams): void => {
-  params.socket.on('send-sideshift-coins-info', (coins: AltpaymentCoin[]) => {
-    params.setCoins(coins.filter(c => c.coin !== params.addressType))
-  })
-  params.socket.on('shift-creation-error', (error: AltpaymentError) => {
-    params.setAltpaymentError(error)
-    params.setLoadingShift(false)
-    return
-  });
-  params.socket.on('quote-creation-error', (error: AltpaymentError) => {
-    params.setAltpaymentError(error)
-    params.setLoadingShift(false)
-    return
-  });
-  params.socket.on('shift-created', (shift: AltpaymentPayment) => {
-    params.setAltpaymentShift(shift)
-    params.setLoadingShift(false)
-  });
-  params.socket.on('send-sideshift-rate', (pair: AltpaymentPair) => {
-    params.setCoinPair(pair)
-    params.setLoadingPair(false)
-  })
-};
+export function getAltpaymentClient (): AltpaymentClient {
+  switch (config.altpaymentClient) {
+    case 'sideshift' as AltpaymentClientOptions:
+      return new SideshiftClient()
+    case 'mocked' as AltpaymentClientOptions:
+      return new MockedPaymentClient()
+    default:
+      throw new Error("ERROR: Invalid alternative payment client")
+  }
+}
