+++ react/lib/config.json
@@ -1,4 +1,5 @@
 {
   "wsBaseUrl": "http://localhost:5000",
-  "apiBaseUrl": "http://localhost:3000"
+  "apiBaseUrl": "http://localhost:3000",
+  "altpaymentClient": "mocked"
 }
