+++ react/lib/util/socket.ts
@@ -1,4 +1,5 @@
 import { Socket } from 'socket.io-client';
+import { AltpaymentCoin, AltpaymentError, AltpaymentPair, AltpaymentPayment } from '../altpayment';
 
 import { BroadcastTxData } from './types';
 
@@ -14,4 +15,39 @@ export const txsListener = (socket: Socket, setNewTxs: Function): void => {
       setNewTxs(unconfirmedTxs);
     }
   });
+};
+
+interface AltpaymentListenerParams {
+  addressType: string
+  socket: Socket
+  setCoins: Function
+  setCoinPair: Function
+  setLoadingPair: Function
+  setAltpaymentShift: Function
+  setLoadingShift: Function
+  setAltpaymentError: Function
+}
+
+export const altpaymentListener = (params: AltpaymentListenerParams): void => {
+  params.socket.on('send-sideshift-coins-info', (coins: AltpaymentCoin[]) => {
+    params.setCoins(coins.filter(c => c.coin !== params.addressType))
+  })
+  params.socket.on('shift-creation-error', (error: AltpaymentError) => {
+    params.setAltpaymentError(error)
+    params.setLoadingShift(false)
+    return
+  });
+  params.socket.on('quote-creation-error', (error: AltpaymentError) => {
+    params.setAltpaymentError(error)
+    params.setLoadingShift(false)
+    return
+  });
+  params.socket.on('shift-created', (shift: AltpaymentPayment) => {
+    params.setAltpaymentShift(shift)
+    params.setLoadingShift(false)
+  });
+  params.socket.on('send-sideshift-rate', (pair: AltpaymentPair) => {
+    params.setCoinPair(pair)
+    params.setLoadingPair(false)
+  })
 };
